# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Backend API", "version": "0.1.0"}, "paths": {"/health": {"get": {"tags": ["health"], "summary": "Root Health", "operationId": "root_health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Root Health Health Get"}}}}}}}, "/api/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "operationId": "register_api_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/me": {"get": {"tags": ["auth"], "summary": "Read Me", "operationId": "read_me_api_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/vaults": {"get": {"tags": ["vaults"], "summary": "List Vaults", "operationId": "list_vaults_api_vaults_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/VaultWithNotes"}, "type": "array", "title": "Response List Vaults Api Vaults Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["vaults"], "summary": "Create Vault", "operationId": "create_vault_api_vaults_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/VaultCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VaultWithNotes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/vaults/{vault_id}": {"get": {"tags": ["vaults"], "summary": "Read Vault", "operationId": "read_vault_api_vaults__vault_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VaultWithNotes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["vaults"], "summary": "Update Vault", "operationId": "update_vault_api_vaults__vault_id__patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VaultUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VaultWithNotes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/vaults/{vault_id}/notes": {"get": {"tags": ["vaults"], "summary": "List Notes", "operationId": "list_notes_api_vaults__vault_id__notes_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/NoteRead"}, "title": "Response List Notes Api Vaults  Vault Id  Notes Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["vaults"], "summary": "Create Note", "operationId": "create_note_api_vaults__vault_id__notes_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NoteCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NoteRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/vaults/{vault_id}/notes/{note_id}": {"patch": {"tags": ["vaults"], "summary": "Update Note", "operationId": "update_note_api_vaults__vault_id__notes__note_id__patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}, {"name": "note_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Note Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NoteUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NoteRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["vaults"], "summary": "Delete Note", "operationId": "delete_note_api_vaults__vault_id__notes__note_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "vault_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Vault Id"}}, {"name": "note_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Note Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "NoteCreate": {"properties": {"title": {"type": "string", "title": "Title", "default": ""}, "content": {"type": "string", "title": "Content", "default": ""}, "links": {"items": {"type": "string"}, "type": "array", "title": "Links"}}, "type": "object", "title": "NoteCreate"}, "NoteRead": {"properties": {"title": {"type": "string", "title": "Title", "default": ""}, "content": {"type": "string", "title": "Content", "default": ""}, "links": {"items": {"type": "string"}, "type": "array", "title": "Links"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "vault_id": {"type": "string", "format": "uuid", "title": "Vault Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "vault_id", "created_at", "updated_at"], "title": "NoteRead"}, "NoteUpdate": {"properties": {"title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}, "content": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Content"}, "links": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Links"}}, "type": "object", "title": "NoteUpdate"}, "RegisterRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}, "display_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Display Name"}}, "type": "object", "required": ["email", "password"], "title": "RegisterRequest"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}, "user": {"$ref": "#/components/schemas/UserRead"}}, "type": "object", "required": ["access_token", "user"], "title": "TokenResponse"}, "UserRead": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "email": {"type": "string", "format": "email", "title": "Email"}, "display_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Display Name"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "email", "created_at", "updated_at"], "title": "UserRead"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "VaultCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "theme": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Theme"}}, "type": "object", "required": ["name"], "title": "VaultCreate"}, "VaultUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "theme": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Theme"}}, "type": "object", "title": "VaultUpdate"}, "VaultWithNotes": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "theme": {"type": "string", "title": "Theme"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "notes": {"items": {"$ref": "#/components/schemas/NoteRead"}, "type": "array", "title": "Notes"}}, "type": "object", "required": ["id", "name", "theme", "created_at", "updated_at"], "title": "VaultWithNotes"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/auth/login"}}}}}}